+++ mods/Farming/mod/data-updates.lua
@@ -1,3 +1,5 @@
 
 require("prototypes.tree_entities-updates")
 require("prototypes.recipe.recipe-updates")
+
+require("prototypes.item.compost_values-updates")