+++ mods/Farming/mod/data.lua
@@ -2,8 +2,11 @@ require("prototypes.item.items")
 require("prototypes.group")
 require("prototypes.recipe_categories")
 
+require("prototypes.entity.compost_bin_entities")
+
 require("prototypes.recipe.root_recipe")
 require("prototypes.recipe.logs_recipe")
 
+require("prototypes.compost_prototypes")
 require("prototypes.patch_prototypes")