+++ mods/Farming/mod/prototypes/group.lua
@@ -9,38 +9,38 @@ data:extend({
   },
   {
     type = "item-subgroup",
-    name = "rs-farming-patch",
-    order = "rs-10-production-machine-004-farming-patch",
-    group = "rs-farming"
-  },
-  {
-    type = "item-subgroup",
-    name = "rs-farming-compost",
-    order = "rs-002-farming-001-farm-compost",
-    group = "rs-farming"
-  },
-  {
-    type = "item-subgroup",
-    name = "rs-log-farming",
-    order = "rs-001-woodcutting-001-logs-farming",
-    group = "rs-farming"
-  },
-  {
-    type = "item-subgroup",
     name = "rs-farming-seeds",
-    order = "rs-002-farming-003-seeds",
+    order = "rs-001-farming-003-seeds",
     group = "rs-farming"
   },
   {
     type = "item-subgroup",
-    name = "rs-farming-roots",
-    order = "rs-002-farming-004-roots",
+    name = "rs-compost",
+    order = "rs-002-farming-001-farm-compost",
     group = "rs-farming"
   },
   {
     type = "item-subgroup",
     name = "rs-compost-recipes",
     order = "rs-002-farming-999-compost-recipes",
+    group = "rs-farming"
+  },
+  {
+    type = "item-subgroup",
+    name = "rs-farming-patch",
+    order = "rs-010-production-machine-004-farming-patch",
+    group = "rs-farming"
+  },
+  {
+    type = "item-subgroup",
+    name = "rs-farming-roots",
+    order = "rs-021-farming-004-roots",
+    group = "rs-farming"
+  },
+  {
+    type = "item-subgroup",
+    name = "rs-log-farming",
+    order = "rs-022-woodcutting-001-logs-farming",
     group = "rs-farming"
   }
 })