+++ mods/Farming/CHANGELOG.md
@@ -1,3 +1,9 @@
+## [0.1.0] - 2019-07-30 - LeonSkills
+### Added
+- Added oak, willow, maple, yew and magic seeds, roots and patches
+- Within the patches you can grow logs (Requires RsWoodcutting) and roots
+- Added compost and compost bins. Some items can now be turned into compost, those items act as "compost fuel". Compost and supercompost increase yield on patches.
+
 ## [0.1.0] - 2019-07-26 - LeonSkills
 ### Added
 - Farming mod created